Replace _wordvec.c with the old 32-bit version from pclu-rh.
The CHANGES file has this entry:

  _wordvec.c: modified for 64-bit architecture

but then a bit later on it has:

  changed definition of CLUREFSZ from 8 back to 4 in lib/gcd.equ

The cumulative effect was that the compiler as shipped in pclu-debian
writes (gcd_tab) and reads (gc_read) lib files with word accesses of
different size.  Fortunately swithcing to 32-bit wordvec is enough to
get them back in sync and fortunately 32-bit wordvec can be fetched
from a previous release.