Do not get lost in the maze of symlinks.
To compile the C files of the newly compiled compiler cd ../code/cmp
which is also the value of bindir in the xfile, no to the ../cmp
symlink (the latter confuses the hell out of furhter ../ traversals).
While here, use subshell to isolate the effects of cd and guard it
with && properly.