Re-bootstrap the compiler.
Recompile the cmpclu compiler with the compiler compiled from the
original bootstrap C sources to get stage 1 compiler, and then
recompile it again with stage1 to get stage 2.  Stage 2 is identical
to stage 1.  Commit these C sources as the new bootstrap.

Most of the differences seems to be purely cosmetic.