Recompile sys/clu with the new compiler.
The diff is trivial.  It's an empty line only diff in all files,
except a few files where proc own variables used to be "static" and
now no longer are.  These is in line with other files, I guess those
few files were not recompiled some time in the past when the compiler
was changed to not emit that "static".