Sync two copies of basic.types
The version in code/cmp/basic.types is the up-to-date one.  Since in
the original distribution cmp was a symlink to code/cmp that's what
the compiler would read at runtime.

Delete the "version control" comment:

  % 3/17/93 dwc: removed stream$_reset (doesn't exist), added stream$scripts

and really delete stream$_reset that was just commented out.

Update the copy in cmpclu to be the same.

Eventually the copy in code/cmp should be gone and the copy in cmpclu
should be installed in cmp (that should not be a symlink to the C
sources).  But small steps for now...