Guard CLU_n definitions with CLU_CLUREF_SMALL_INTEGERS.
The code used to define CLU_n as CLUREF objects when #ifdef sparc.
What it really means is as follows, as far as I can tell.  IIRC, the
old Sun compiler (hence "sparc") passed all structures by reference,
while gcc passes small structures "as an integer".  CLUREF is a small
structure, so when you pass an integer to a function without a
prototype, that integer is passed just like a CLUREF that contains
that integer would be.  Make that explicit (and test that it actually
works).