Use C99 compound literals for CLU_n small integers.
Do not leak ABI details into the source code and don't rely on
aliasing behaviour (that passing an integer is the same at the ABI
level as passing an integer-sized struct).

On smaller test functions this kind of change results in the same
binary code but on larger ones the batterfly effect propagates through
the compiler and it's not quite the same instruction for instruction.