sysasm: util.c - add prototypes for the CLU OPs used.
With CLU_n definitions fixed it requires only one cast to take care of
aliasing: a pointer to CLUSTRING (&foo_STRING messages) is also a
pointer to a CLUREF that contains that CLUSTRING.

Cheat for now and declare stringOPcons to have buf as const char *
It's C_STRING in the definition, but is de-facto const.  Adding const
correctness is a whole different yak to shave.