sysasm: connected_dir.c - cleanup
ANSIfy and fix warnings.  Prettify.
G/c RCS Header and Log.

#ifdef out the "probably unnecessary" code to save the error message
in elist[1], but fix it to do the right thing: getwd returns the error
message in the buffer on error - that's why the error check was after,
the stringOPcons.  getcwd doesn't do that, so adjust accordingly.

Don't add an extra trailing slash when the pwd is the root directory.