current_apply$is_type_own_init - a meaningful name for a cryptic test
An operation with an empty name in the current type represents, as far
as I can tell, the own init for that type.  is_type_own_init is
intended to be a more comprehensible alternative for the equivalent
open-coded test of that condition:

    current_type$exists() cand
        current_apply$get_name() = current_type$get_name()

that gets even more cryptic when this is just one term in a more
complex condition.

Use the new op in g_ownvars to make it explicit which code is
generated when.  I was trying to fix the indentation of the code it
emits but got lost and had to backtrack.

No functional change is intended.