g_owninit - fix/modernize indentation of the emitted C code.
This is mostly trivial re-shuffling of indent/outdent lines, except
for the trailer.  The trailier used to emitt the closing brace in two
places under different conditions, which is confusing.  Reorganize
that code so that we only emit it in one place under the same
condition the opening one was emitted under.  As far as I can tell,
this doesn't actually change the effective condition, only the textual
expression.