Define struct with local vars once, refer it in iterator bodies.
Split g_localvars into several functions and recombine them so that we
can define struct for the local separately if needed.

G/c current_locals and ig$gen_locals (that repeated the work of
defining the struct in the argument list of iterator bodies) as they
are no longer needed.