g_iterbody_names, g_iterbody_names_asst - clarify
"arr" and "out" are not very descriptive names, so try to be a bit
more readable.  While here, move the interesting tagcase (for_) to the
front and add a few comments to make it stand out a bit more from the
"recurse into nested statements" boilerplate.

Same C code is generated.