errcmp - cast strcmp() argument, add a FIXME comment.
FIXME: Compiler only emits an errcmp() call when s2 is a string (it
emits == otherwise), hence the first "false".  But then checking s2
for being a numeric error is redundant.  Without that context the
definition of errcmp() seems self-contradictory at first glance.
Supply that context in a comment for now and revisit later.