gdefn.clu - void *user_locals to dodge warnings.
char ** doesn't make much sense in general and causes problems when
you define an iterator first and then use it later (stable$entries in
the compiler, for example).  Often that warning is avoided b/c the
compiler doesn't emit prototypes for external functions, but if you
define the iterator first, its definition provides one and you do get
the warning.