g_own_build_one - cast to OPSPTR ops passed to add_parm_info_type.
We pass pointers to structures that are compatible.  Unfortunately
with pedantic C99 you cannot use a structure with a flexible member as
a header in another structure, so dressing this up in legal C is not
immediately straightforward.

This likely needs to be applied to other cases where we emit
add_parm_info_type() calls, but I'd like to actually look at them
first, they are not used in the compiler code itself.

The compiler output of itself now compiles without warnings with
-Wall -Wextra -Wno-unused-*