g_pownvars_init - must emit the second own_init test even if there're no owns
This is a regression introduced in ee23589cc836 refactoring ("While
here, don't emit own init block if there's nothing to init.").  The
own_init flag is also used to init g_tbl/g_const earlier, but if there
were no owns we would not emit the second check and so wouldn't set
the flag to indicate the init was done.  Always emit the second check
and let the C compiler merge the two ifs if it's possible.