Regen committed C sources with the current compiler.
It's long overdue, but it creates huge diffs and the compiler could
still be bootstrapped (I think :) with the old C sources.  But the
forthcoming change to g/c CLU_string_size temporary variable that the
new compiler no longer uses will definitely break it, so commit the
sources that don't use it (and are mostly warning free too as a bonus).