pclu_sys.h - Wrap resignal() in do/while(0).
While here, test for ERR_ok first to make it more visible.

The name of this macro is misleading.  Clu's "resignal" statement
re-signals the original error, but the macro always wraps an error
into an ERR_failure, like Clu's implicit failure propagation does.
Also returning on ERR_ok is a bit surprising and the "asm" C sources
almost always do if (err != ERR_ok) resignal(err);

Long term it's probably better to use the same patter the compiler
itself uses in the generated C code with the single "ex_0:" block at
the end of the function that does the implicit ERR_failure wrapping.