sysasm: string.c - Simplify string$chars().
I'm not sure what the point of the explicit list of

  if (err == ERR_iterfoo) signal(err);

exactly is.  Code emitted by the compiler also has a catch all
signal(err) at the end of that sequence that was missing here.  This
is all equivalent to just a plain signal(err) as far as I can tell.