sysasm: Comment debug_print -> print forwarding for selectors.
Note that we abuse^W rely on the fact that print and debug_print have
similar op_owns, so debug_print calls print but with debug_print field
ops.  Should probably just do a tail call.

Add this comment to the already cleaned-up oneof, variant and struct.
Not added to record.c as it's not cleaned up yet, but it does already
have this fact briefly mentioned.