libasm: _ioctl.c - ANSIfy.  Fix up a bit, but it's still broken.
It's not entirely clear to me how this is supposed to be used.

The original obviously never worked anyway b/c it used H << 16 + L
to reassemble the 32-bit request, which means H << (16 + L) - not
(H << 16) + L it intends to say.

Also this file doesn't provide _ioctl_of_t_reqs and _ioctl_ownreqs,
so the code calling _ioctl[]() just wouldn't link.