sysasm: get_jcl.c - Fix regression introduced in fa0662a5fb4c.
Original code always added a trailing space.  The +1 in the size
calculations was confusing b/c I misinterpreted it as "space or NUL",
but the string implementation itself allocates an extra byte for NUL
behind the scenes and it need not be explicitly requested.

An alternative fix would be to decrement size.  But restore the
original behavior in case something might depend on that final space
being there.