g_int_const - Don't emit non-existent CLU_100500 constants.
Introduce int_const$g_emit that emits CLU_n if it exists or
CLUREF_make_num(n) otherwise.

g_int_const is used for instantiation of parmd modules, so bad code
was generated for large int parameters (sometype[100500]).

We can always emit the make_num form (CLU_n are defined that way
anyway), but minimize the churn in the generated C code for now.
Nothing in the compiler itself hits this problem.

The old code that emits a temp variable can be g/c'ed but it's a bit
of a rototill b/c this percolates through the code as the temps has to
be declared etc.