sysasm: new_parm_util.c - Try to name things more consistently.
There are too many different kinds of "ops" floating around, so try to
clarify what is what.  There are three kinds:

"aops" - abstract ops, a template.  This is a global variable that the
compiler emits for a parmd type.  It has a dummy for owns.

"instance" - instantiation of an "aops".  Allocated dynamically.  What
gives it its identity are the "owns", where the parmd code keeps its
clu "own" variables and gets its interface to the specific parms it's
called with.

"table" - this is a vtable of sorts.  While aops/instance opsptr
corresponds to a cluster, "table" opsptr corresponds to the interface
that the parmd code expects of its parm.  It's constructed based on
parmd code REQS.