g_parm_build_one - Try to clarify a bit.
"yes" is not the most informative variable name, especially when one
of the reasons to set it seems obsolete.  The test that checks it puts
the is_op check first, making it stand out, and then, confusingly,
proceeds to the code to instantiate a type.

Rename s/yes/is_type/ (not the best name either, but still much
better) and add some comments.  This can be improved further, but I'm
not enitrely sure in the details, so leave it at that for now.