g_def_opsptr - Factor out definition of opsptr instances.
Get the "pass" parameter from gtbl$g_field_tables() all the way down
here - if it's "static", the opsptr we are defining is a top-level
global variable.  In that case declare it as "common".

These opsptrs are instantiations of parmd types or selectors and they
are defined in each file that uses corresponding clu type, so we do
need them to be common.