Work around a problem with (private!) GC_jmp_buf in gc-7.x
Some "asm" C sources peek under the hood of GC and include gc_priv.h.
Unfortunately in gc-7.x that header defines GC_jmp_buf (gc-8.x fixed
this and just declares it).  Play some cpp tricks to give that
variable a new name in each file that does this, so that if we compile
with -fno-common we don't get multiply defined symbol.