Move the files compiler reads at runtime to ~CLU/cmp
In the old native compiler distributions for vax and sun the directory
~CLU/cmp used to contain the compiler sources in clu and asm.

In the portable clu distribution the CLU sources were moved to
~CLU/clucmp and the C output from pclu now goes to code/cmp with
toplevel ~CLU/cmp made a symlink to these pregenerated C sources.
So this created a bit of a maze.

The compiler still reads basic.types from the old location and also
reads pclu-specific cc*.cmd files from there.  Untangle the maze by
placing them directly where they are expected to be, now that ~CLU/cmp
is a real directory, not a symlink.