g_proc_exprs - clarify the clusterop/idn "skip" case.
The original comment was cryptic and not very informative.  What's
really going on here is that we are dealing with "where" restrictions
on type parms.  These functions are not called directly, so there's no
extern dependecy to declare.  While here rename the flag variable to
better reflect its meaning and change the orider of checks to be more
natural (parm$member_op is additional restrictions).