g_proc_exprs - add a comment about the op/apply case.
As this is named somewhat confusingly, it's not immediately obvious
that "apply" here refers to the internal "applytype" which is a
meta-clus (if you excuse the pun) for "proctype" and "itertype".
Cf. the specs in cmp/basic.types and implementations in sysasm.

To quote clu.hacks (that was _not_ shipped as part of pclu and was
fetched from native CLU compilers):

  The keyword [...] APPLYTYPE can be used in place of CLUSTER.  [...]
  APPLYTYPE module exists only for the purpose of writing (in CLU) the
  specifications of the proctype and itertype generators [...] as
  such, they can be type-checked, but cannot be compiled.