build_one_selector_table - cast to OPSPTR ops passed to add_selector_info.
This is a counterpart to revision 90e8cc4e2ed7 (add_parm_info_type).

We pass pointers to structures that are compatible.  Unfortunately
with pedantic C99 you cannot use a structure with a flexible member as
a header in another structure, so dressing this up in legal C is not
immediately straightforward.