G/c include/debug/pclu_err.h
Original source tree has a bunch of symlinks in include/debug and an
override for include/debug/pclu_err.h that defines CLU_DEBUG to 1 and
includes the real one.

We now just provide that define on the compiler invocation and can g/c
the last remnant of the old include/debug.