sysasm: G/c debug/util.c, use the one from sysasm.
In the most recent pclu-debian distribution debug/util.c and
sysasm/Opt/util.c were practically identical modulo trivial
sys_errlist vs strerror change.

In the more older pclu-LINUX-RH7 distribution the essential change
between the two versions was the call to debug$cli(GENESIS, ...)
before the call to start_up.  Just fold that small bit of code into
the sysasm version under an ifdef.

That version also had that GENESIS + start_up block in an infinite
loop.  It's nice to keep the breakpoints and whatever and be able to
re-run the program, but since we are in the same old process, the
global state is modified by the previous run, so drop the loop for
now.