gtypes.clu - rename externalize to s_externalize (kludge)
The debugger has the same externalize proc and since the debugger is
linked directly into the program there's a name conflict.

Fixing it here is easier since there's only one caller and s_ prefix
is actually in line with the naming convention (s_foo returns a
string, g_foo emits the string).