plink - pass as much args as we can to cc/ld
Instead of manually remapping plink args to cc/ld args (like -o or
-prof) which is necessarily limited, just pass about everything we can
to cc/ld.  The only options we process are -opt/-debug and -verbose.
The first non-option argument, an unknown option or explicit -- makes
the rest of them passed to cc/ld.

This makes it possible to use third-party libraries and/or fancy
linking options by just using them "normally" on the plink line.  As
an additional bonus, filenames with spaces in them should be now
handled correctly too.