gdbg.clu - start prettifying emitted popowns
Add some newlines and indentation.

Needs more rework to not pass the "first" flag around.  Don't need it
b/c C allows trailing comma, so we can always emit one after an
initializer instead of keeping track of whether we need to emit one
before.  But that's a more intrusive rototill, so leave it for later.