dbg_info$emit_decl - fix dbg_info::count value in the initializer
I don't think this has been used in ages, but make it match the
definition that has 14 intptr words, not 12.  My guess is that at some
point ptowns and popowns were forked off of owns, adding two more
words, but the change was not reflected in the count, and it was not
noticed b/c the count is not actually used for anything.