gdefn.clu - add a comment about wrong dbg_info$start behavior
g_applydefn calls dbg_info$start.  It calls emit_own_decl internally,
but for the first op of a type we have just stuffed the owns list
(dbg_info$add_ilist) with inits for *all* the ops out of context.  So
e.g. for ce.clu we end up with a vlist of 90+ "init" owns that all get
the name from the first op.  Some of those inits are bogus too since
not every op has owns.

This requires non-trivial rework.