g_proc_exprs - add a comment about a_cons case
array cons is syntactic sugar for array$create or array$new plus a
series of array$addh calls, so we should add those names to the list
too.  pclu_sys.h happens to already declare new and addh, but not
create, so for some sources you can see a warning for the latter.
It should probably pre-declare all of them and be done with it.