Make error messages more unixy.
This helps integration with existing tools that expect

  file.clu:42: rtfm

message format (emacs compilation-mode, etc).

Hopefully this is not too kludgy, though the original sources where a
bit rough to begin with.  Mostly it's a question of dragging the
unmodified soruce name through a few function calls and tweaking the
output format a bit to include it.