g_resignal - prettify generated code
Don't lump the "if" and its then clause on the same line.  If you want
to debug a Clu program with a C debugger, you probably want to be able
to set a break point inside that then clause.

Don't repeat the name of the signal, just signal the value of "err"
(that we know).  The iterator body case did that already (for iecode).