g_yield - one unconditional signal() is enough to propagate all errors
The code used to propagate break, iterbodyreturn, iterbodyexit and
iterbodysignal separately.  The "revision control" comments above this
block of code seem to suggest that different calling convention used
to be used for them.  But now the unconditional signal(err) at the end
will also propagate them just as well, so drop the now unnecessary
individual checks for them.