g_tagcase - rewrite from scratch
The old version emitted code that was suboptimal and arguably not
correct.  For a tag arm with multiple tags the old code would
(suboptimally) evaluate all is_tag checks without short-circuiting and
then (incorrectly) call value_tag of the first tag to get the value.

Rewrite from scratch to test tags from left to right with proper
short-circuiting.  Use the value accessor that corresponds to the
matched tag.