pclu_sys.h - bring back oneofOPnew, add arrayOPcreate
"dumb stuff for alpha" is ops emitted by the compiler for array cons.
Make them proper prototypes and move them with the rest of the
prototypes in the next section.  Add arrayOPcreate and bring back
oneofOPnew that a lot of _gcd native ops use.

All this is a bit ad-hoc, but it will do for now.