g_qop - fix indentation of emitted overflow checks
I'm not sure if that extra column indentation for the test was
intentional.  If so (like, say, perl's "or die" indentation), it is
too small IMHO: with just one column it's enough to be noticable, but
not enough to clearly stand apart from the "main" flow.  So instead of
getting out of the way it attracts unwanted attention.