g_expr0 - don't declare a temp for array size for a_cons.
This augments 13b3e5ff709f (g_arraycons - don't emit unused size_nm
temp).  The code removed in that revision did not emit anything, but
was just skipping over the name it was not going to use, the
declaration actually happens here in g_expr0 ("spooky action at a
distance").