pclu(1): mention that type library can be dumped
The original language in pclu.tex might create an impression that the
"type library" it talks about is a something external and permanent.
Try to make it more clear that it's the in-core type library that is
maintained only during a signle compiler run.  While here, it's a good
place to mention the library indeed can be dumped for later use.